Construa os testes relativos ao código - part 1

+++ src/main/java/com/eCommerce/VirtualBookStore/model/entities/OrderItem.java
@@ -27,7 +27,7 @@ public class OrderItem {
     public OrderItem(@NotNull Book book, @Positive Integer amount) {
         this.book = book;
         this.amount = amount;
-        this.momentPrice = BigDecimal.valueOf(book.getPrice());
+        this.momentPrice = book.getPrice();
     }
 
     public Book getBook() {
